requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Entity Resolution
splink==4.0.0
duckdb==0.10.0

# Data Processing
pandas==2.2.0
pyarrow==15.0.0
openpyxl==3.1.2
python-multipart==0.0.9
lxml==5.1.0

# Task Queue (Phase 2)
celery[redis]==5.3.6
redis==5.0.1

# Authentication (Phase 3)
authlib==1.3.0
httpx==0.26.0
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.1
faker==22.6.0
tqdm==4.66.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
